refactor: split calendar component
